Make SingleVector and AsUninit #[repr(transparent)].
This to allow for a future AsRefs trait to be implemented for these
types. That would be especially useful for I/O vectors, where code may
want the sematics of using such wrapper types, but still be layout
compatible with the way the OS reads the I/O vectors of a given system
call.